reformat code
